reworked models added testing jig
All test-bed modules have been reworked to be more 'printer friendly'.
Print 90deg on the flat side of the unit_side connector for easier
printing. Still a difficult print, but a tad less so.

Testing jig has been added for magnet strength. This will be heavily
revised this is only for light initial shock testing.